[Refactored] Cleaned naming code to its own class
